FEED-100: Add an option to send a reminder for all open feedback older than a week.
It would be nice if this was called automatically on a daily basis in the future.
