+++ backend/feedbag/organization/admin.py
@@ -11,19 +11,24 @@ class OrganizationAdmin(DjangoObjectActions, admin.ModelAdmin):
     def import_users(self, request, obj):
         g = GlassFrogImporter(api_key=obj.glassfrog_api_key, organization=obj)
         g.import_users()
+    import_users.label = _('Import users for this organization')
 
     def import_roles(self, request, obj):
         g = GlassFrogImporter(api_key=obj.glassfrog_api_key, organization=obj)
         g.import_circles(anchor_circle_id=obj.glassfrog_anchor_circle_id)
+    import_roles.label = _('Import roles for this organization')
 
     def archive_roles(self, request, obj):
         obj.roles.update(archived=True)
-
-    import_users.label = _('Import users for this organization')
-    import_roles.label = _('Import roles for this organization')
     archive_roles.label = _('Archive existing roles for this organization')
 
-    change_actions = ('import_users', 'import_roles', 'archive_roles')
+    def send_reminder(self, request, obj):
+        obj.message_for_reminder()
+
+    send_reminder.label = _('Send reminder')
+    send_reminder.short_description = _('Send a reminder to users that have unfinished feedback older than a week.')
+
+    change_actions = ('import_users', 'import_roles', 'archive_roles', 'send_reminder')
 
 
 admin.site.register(Organization, OrganizationAdmin)
